create module for get titled players
